Feat: Api연결, 기능조정

See merge request s09-final/S09P31B109!159

+++ FrontEnd_App/bangrang/app/src/main/java/com/ssafyb109/bangrang/view/utill/RankProfile.kt
@@ -26,7 +26,7 @@ import coil.request.ImageRequest
 import com.ssafyb109.bangrang.R
 
 @Composable
-fun RankProfile(image: String?, percentage: Int, userId: String, modifier: Modifier = Modifier, rank: Int) {
+fun RankProfile(image: String?, percentage: Double, userId: String, modifier: Modifier = Modifier, rank: Int) {
     val medal = when(rank) {
         1 -> painterResource(id = R.drawable.first)
         2 -> painterResource(id = R.drawable.second)
